apps/.helm/producer/requirements.txt
psycopg2-binary==2.9.5
fastapi==0.86.0
uvicorn==0.19.0
asyncio
statistics
httpx==0.23.1
kafka-python==2.0.2